#112 Add Some rest endpoints that are required in the frontend requirements and improvements views

+++ business/src/main/java/com/kevinguanchedarias/owgejava/util/ValidationUtil.java
@@ -119,7 +119,7 @@ public class ValidationUtil {
 	 * @author Kevin Guanche Darias <kevin@kevinguanchedarias.com>
 	 */
 	public ValidationUtil requirePositiveNumber(Integer target, String position) {
-		return requirePositiveNumber(target.intValue(), position);
+		return requirePositiveNumber(Long.valueOf(target), position);
 	}
 
 	/**
